May addition

+++ src/main/java/basepackage/stand/standbasisprojectonev1/model/Attendance.java
@@ -47,7 +47,7 @@ public class Attendance extends DateAudit{
     @NotNull
     private Integer done; // 0- not done, 1- done, -1 voided, 2 - late
     
-    private Integer delegated; // ID of user that is assigned to use this Calendar
+    private Long delegated; // ID of teacher that is assigned to use this Attendance
     
     @Size(max = 500)
     private String _desc;
